pyproject.toml
[project]
name = "obsidian-tmdb-cover"
version = "0.1.0"
description = "Add TMDB cover images to Obsidian notes"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
keywords = ["obsidian", "tmdb", "movies", "tv", "covers", "metadata"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Office/Business :: Office Suites",
    "Topic :: Text Processing :: Markup",
    "Topic :: Utilities",
]
dependencies = [
    "requests>=2.31.0",
    "pyyaml>=6.0.1",
    "pillow>=10.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/obsidian-tmdb-cover"
Repository = "https://github.com/yourusername/obsidian-tmdb-cover.git"
Issues = "https://github.com/yourusername/obsidian-tmdb-cover/issues"

[project.scripts]
obsidian-cover = "obsidian_tmdb_cover.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
    "types-requests",
    "types-PyYAML",
]